requirements.txt
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
newspaper3k==0.2.8
textblob==0.18.0
Pillow==11.2.1
requests==2.32.3